requirements.txt
bottle==0.13.2
cffi==1.17.1
clr-loader==0.2.6
proxy_tools==0.1.0
pycparser==2.22
pythonnet==3.0.4
pywebview==5.3.2
typing_extensions==4.12.2
